feat: creating plans & agents modules
