+++ src/a2a/server.ts
@@ -74,7 +74,7 @@ export interface PaymentsA2AServerResult {
  * @param res - Express response object
  * @param next - Express next function
  */
-function bearerTokenMiddleware(
+function _bearerTokenMiddleware(
   handler: PaymentsRequestHandler,
   req: express.Request,
   res: express.Response,
@@ -244,7 +244,7 @@ export class PaymentsA2AServer {
     }
 
     if (exposeDefaultRoutes) {
-      appBuilder.setupRoutes(app, basePath)//, [bearerTokenMiddleware.bind(null, handler)])
+      appBuilder.setupRoutes(app, basePath) //, [bearerTokenMiddleware.bind(null, handler)])
     }
 
     if (exposeAgentCard) {
