+++ src/common/types.ts
@@ -16,7 +16,7 @@ export interface PaymentOptions {
    * You can get your API key by logging in to the Nevermined App.
    * @see https://docs.nevermined.app/docs/tutorials/integration/nvm-api-keys
    */
-  nvmApiKey?: string
+  nvmApiKey: string
 
   /**
    * The URL to return to the app after a successful login.
