+++ src/payments.ts
@@ -1,41 +1,22 @@
-import { decodeJwt } from 'jose'
 import { AIQueryApi } from './api/query-api'
-import { jsonReplacer } from './common/helper'
 import { PaymentsError } from './common/payments.error'
 import { decodeAccessToken } from './utils'
 import {
   PaymentOptions,
-  PlanPriceConfig,
-  PlanCreditsConfig,
-  PlanCreditsType,
-  AgentMetadata,
-  AgentAPIAttributes,
-  PlanBalance,
-  PlanMetadata,
   AgentAccessParams,
   ValidationAgentRequest,
-  PaginationOptions,
   TrackAgentSubTaskResponseDto,
   TrackAgentSubTaskDto,
   NvmAPIResult,
   StripeCheckoutResult,
 } from './common/types'
-import { EnvironmentInfo, Environments } from './environments'
-import { getRandomBigInt, isEthereumAddress } from './utils'
 import {
   API_URL_ADD_PLAN_AGENT,
   API_URL_BURN_PLAN,
-  API_URL_GET_AGENT,
   API_URL_GET_AGENT_ACCESS_TOKEN,
-  API_URL_GET_AGENT_PLANS,
-  API_URL_GET_PLAN,
-  API_URL_GET_PLAN_AGENTS,
   API_URL_MINT_EXPIRABLE_PLAN,
   API_URL_MINT_PLAN,
   API_URL_ORDER_PLAN,
-  API_URL_PLAN_BALANCE,
-  API_URL_REGISTER_AGENT,
-  API_URL_REGISTER_PLAN,
   API_URL_REMOVE_PLAN_AGENT,
   API_URL_STRIPE_CHECKOUT,
   API_URL_TRACK_AGENT_SUB_TASK,
@@ -44,25 +25,26 @@ import {
 } from './api/nvm-api'
 import * as a2aModule from './a2a'
 import type { PaymentsA2AServerOptions, PaymentsA2AServerResult } from './a2a/server'
+import { BasePaymentsAPI } from './api/base-payments'
+import { PlansAPI } from './api/plans-api'
+import { AgentsAPI } from './api/agents-api'
 
 /**
  * Main class that interacts with the Nevermined payments API.
  * Use `Payments.getInstance` for server-side usage or `Payments.getBrowserInstance` for browser usage.
  * @remarks This API requires a Nevermined API Key, which can be obtained by logging in to the Nevermined App.
  */
-export class Payments {
+export class Payments extends BasePaymentsAPI {
   public query!: AIQueryApi
-  public returnUrl: string
-  public environment: EnvironmentInfo
-  public appId?: string
-  public version?: string
-  public accountAddress?: string
-  private nvmApiKey?: string
-  public isBrowserInstance = true
+  public plans!: PlansAPI
+  public agents!: AgentsAPI
 
   /**
    * Exposes A2A agent/server functionality for this Payments instance.
-   * Usage: payments.a2a.start({ agentCard, executor, port, ... })
+   * @example
+   * ```
+   * payments.a2a.start({ agentCard, executor, port, ... })
+   * ```
    */
   public readonly a2a: {
     /**
@@ -115,8 +97,21 @@ export class Payments {
    */
   static getBrowserInstance(options: PaymentOptions) {
     if (!options.returnUrl) {
-      throw new PaymentsError('nvmApiKey is required')
+      throw new PaymentsError('returnUrl is required')
     }
+    const url = new URL(window.location.href)
+    const urlNvmApiKey = url.searchParams.get('nvmApiKey') as string
+    if (urlNvmApiKey) {
+      url.searchParams.delete('nvmApiKey')
+    }
+
+    const urlAccountAddress = url.searchParams.get('accountAddress') as string
+    if (urlAccountAddress) {
+      url.searchParams.delete('accountAddress')
+    }
+
+    history.replaceState(history.state, '', url.toString())
+
     return new Payments(options, true)
   }
 
@@ -127,16 +122,11 @@ export class Payments {
    * @param isBrowserInstance - Whether this instance is for browser usage.
    */
   private constructor(options: PaymentOptions, isBrowserInstance = true) {
-    this.nvmApiKey = options.nvmApiKey
-    this.returnUrl = options.returnUrl || ''
-    this.environment = Environments[options.environment]
-    this.appId = options.appId
-    this.version = options.version
+    super(options)
+
     this.isBrowserInstance = isBrowserInstance
-    if (!this.isBrowserInstance) {
-      this.parseNvmApiKey()
-      this.initializeApi()
-    }
+    this.parseNvmApiKey()
+    this.initializeApi(options)
     // ---
     // Attach the a2a server API to this instance
     this.a2a = {
@@ -151,22 +141,12 @@ export class Payments {
   }
 
   /**
-   * Parses the NVM API Key to extract the account address.
-   * @throws PaymentsError if the API key is invalid.
-   */
-  private parseNvmApiKey() {
-    try {
-      const jwt = decodeJwt(this.nvmApiKey!)
-      this.accountAddress = jwt.sub
-    } catch (error) {
-      throw new PaymentsError('Invalid NVM API Key')
-    }
-  }
-
-  /**
    * Initializes the AI Query Protocol API.
    */
-  private initializeApi() {
+  private initializeApi(options: PaymentOptions) {
+    this.plans = PlansAPI.getInstance(options)
+    this.agents = AgentsAPI.getInstance(options)
+
     this.query = new AIQueryApi({
       backendHost: this.environment.backend,
       apiKey: this.nvmApiKey!,
@@ -187,50 +167,8 @@ export class Payments {
    */
   public connect() {
     if (!this.isBrowserInstance) return
-    const url = new URL(
-      `/en/login?nvm-export=nvm-api-key&returnUrl=${this.returnUrl}`,
-      this.environment.frontend,
-    )
+    const url = new URL(`/login?returnUrl=${this.returnUrl}`, this.environment.frontend)
     window.location.href = url.toString()
-  }
-
-  /**
-   * Initializes the class after the user has logged in and been redirected
-   * back to the app ({@link returnUrl}).
-   *
-   * @remarks
-   * This is a browser-only function.
-   * @example
-   * ```
-   * payments.init()
-   * ```
-   * @example Using React
-   * ```
-   * useEffect(() => {
-   *   payments.init()
-   * })
-   * ```
-   */
-  public init() {
-    if (!this.isBrowserInstance) return
-    const url = new URL(window.location.href)
-    const nvmApiKey = url.searchParams.get('nvmApiKey') as string
-
-    if (nvmApiKey) {
-      this.nvmApiKey = nvmApiKey as string
-      url.searchParams.delete('nvmApiKey')
-    }
-
-    const accountAddress = url.searchParams.get('accountAddress') as string
-
-    if (accountAddress) {
-      this.accountAddress = accountAddress
-      url.searchParams.delete('accountAddress')
-    }
-
-    history.replaceState(history.state, '', url.toString())
-    this.initializeApi()
-    this.parseNvmApiKey()
   }
 
   /**
@@ -244,7 +182,7 @@ export class Payments {
    * ```
    */
   public logout() {
-    this.nvmApiKey = undefined
+    this.nvmApiKey = ''
   }
 
   /**
@@ -256,414 +194,7 @@ export class Payments {
    * @returns True if the user is logged in.
    */
   get isLoggedIn(): boolean {
-    return !!this.nvmApiKey
-  }
-
-  /**
-   *
-   * It allows to an AI Builder to register a Payment Plan on Nevermined in a flexible manner.
-   * A Payment Plan defines 2 main aspects:
-   *   1. What a subscriber needs to pay to get the plan (i.e. 100 USDC, 5 USD, etc).
-   *   2. What the subscriber gets in return to access the AI agents associated to the plan (i.e. 100 credits, 1 week of usage, etc).
-   *
-   * With Payment Plans, AI Builders control the usage to their AI Agents.
-   *
-   * Every time a user accesses an AI Agent to the Payment Plan, the usage consumes from a capped amount of credits (or when the plan duration expires).
-   * When the user consumes all the credits, the plan automatically expires and the user needs to top up to continue using the service.
-   *
-   * @remarks
-   * This method is oriented to AI Builders.
-   * The NVM API Key must have publication permissions.
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/create-plan
-   *
-   * @param planMetadata - @see {@link PlanMetadata}
-   * @param priceConfig - @see {@link PlanPriceConfig}
-   * @param creditsConfig - @see {@link PlanCreditsConfig}
-   * @param nonce - Optional nonce to prevent replay attacks. Default is a random BigInt.
-   * @example
-   * ```
-   *  const cryptoPriceConfig = getNativeTokenPriceConfig(100n, builderAddress)
-   *  const creditsConfig = getFixedCreditsConfig(100n)
-   *  const { planId } = await payments.registerPlan({ name: 'AI Assistants Plan'}, cryptoPriceConfig, creditsConfig)
-   * ```
-   *
-   * @returns The unique identifier of the plan (Plan ID) of the newly created plan.
-   */
-  public async registerPlan(
-    planMetadata: PlanMetadata,
-    priceConfig: PlanPriceConfig,
-    creditsConfig: PlanCreditsConfig,
-    nonce = getRandomBigInt(),
-  ): Promise<{ planId: string }> {
-    const body = {
-      metadataAttributes: planMetadata,
-      priceConfig,
-      creditsConfig,
-      nonce,
-      isTrialPlan: planMetadata.isTrialPlan || false,
-    }
-    const options = this.getBackendHTTPOptions('POST', body)
-    const url = new URL(API_URL_REGISTER_PLAN, this.environment.backend)
-
-    const response = await fetch(url, options)
-    if (!response.ok) {
-      throw Error(`${response.statusText} - ${await response.text()}`)
-    }
-
-    return response.json()
-  }
-
-  /**
-   *
-   * It allows to an AI Builder to create a Payment Plan on Nevermined based on Credits.
-   * A Nevermined Credits Plan limits the access by the access/usage of the Plan.
-   * With them, AI Builders control the number of requests that can be made to an agent or service.
-   * Every time a user accesses any resouce associated to the Payment Plan, the usage consumes from a capped amount of credits.
-   * When the user consumes all the credits, the plan automatically expires and the user needs to top up to continue using the service.
-   *
-   * @remarks This method is oriented to AI Builders
-   * @remarks To call this method, the NVM API Key must have publication permissions
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/create-plan
-   *
-   * @param planMetadata - @see {@link PlanMetadata}
-   * @param priceConfig - @see {@link PlanPriceConfig}
-   * @param creditsConfig - @see {@link PlanCreditsConfig}
-   *
-   * @example
-   * ```
-   *  const cryptoPriceConfig = getNativeTokenPriceConfig(100n, builderAddress)
-   *  const creditsConfig = getFixedCreditsConfig(100n)
-   *  const { planId } = await payments.registerCreditsPlan({ name: 'AI Credits Plan'}, cryptoPriceConfig, creditsConfig)
-   * ```
-   *
-   * @returns The unique identifier of the plan (Plan ID) of the newly created plan.
-   */
-  public async registerCreditsPlan(
-    planMetadata: PlanMetadata,
-    priceConfig: PlanPriceConfig,
-    creditsConfig: PlanCreditsConfig,
-  ): Promise<{ planId: string }> {
-    if (
-      creditsConfig.creditsType != PlanCreditsType.FIXED &&
-      creditsConfig.creditsType != PlanCreditsType.DYNAMIC
-    )
-      throw new PaymentsError('The creditsConfig.creditsType must be FIXED or DYNAMIC')
-
-    if (creditsConfig.minAmount > creditsConfig.maxAmount)
-      throw new PaymentsError(
-        'The creditsConfig.minAmount can not be more than creditsConfig.maxAmount',
-      )
-
-    return this.registerPlan(planMetadata, priceConfig, creditsConfig)
-  }
-
-  /**
-   *
-   * It allows to an AI Builder to create a Payment Plan on Nevermined limited by duration.
-   * A Nevermined Credits Plan limits the access by the access/usage of the Plan.
-   * With them, AI Builders control the number of requests that can be made to an agent or service.
-   * Every time a user accesses any resouce associated to the Payment Plan, the usage consumes from a capped amount of credits.
-   * When the user consumes all the credits, the plan automatically expires and the user needs to top up to continue using the service.
-   *
-   * @remarks This method is oriented to AI Builders
-   * @remarks To call this method, the NVM API Key must have publication permissions
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/create-plan
-   *
-   * @param planMetadata - @see {@link PlanMetadata}
-   * @param priceConfig - @see {@link PlanPriceConfig}
-   * @param creditsConfig - @see {@link PlanCreditsConfig}
-   *
-   * @example
-   * ```
-   *  const cryptoPriceConfig = getNativeTokenPriceConfig(100n, builderAddress)
-   *  const 1dayDurationPlan = getExpirableDurationConfig(ONE_DAY_DURATION)
-   *  const { planId } = await payments.registerTimePlan({ name: 'Just for today plan'}, cryptoPriceConfig, 1dayDurationPlan)
-   * ```
-   *
-   * @returns The unique identifier of the plan (Plan ID) of the newly created plan.
-   */
-  public async registerTimePlan(
-    planMetadata: PlanMetadata,
-    priceConfig: PlanPriceConfig,
-    creditsConfig: PlanCreditsConfig,
-  ): Promise<{ planId: string }> {
-    if (creditsConfig.creditsType != PlanCreditsType.EXPIRABLE)
-      throw new PaymentsError('The creditsConfig.creditsType must be EXPIRABLE')
-
-    return this.registerPlan(planMetadata, priceConfig, creditsConfig)
-  }
-
-  /**
-   *
-   * It allows to an AI Builder to create a Trial Payment Plan on Nevermined limited by duration.
-   * A Nevermined Trial Plan allow subscribers of that plan to test the Agents associated to it.
-   * A Trial plan is a plan that only can be purchased once by a user.
-   * Trial plans, as regular plans, can be limited by duration (i.e 1 week of usage) or by credits (i.e 100 credits to use the agent).
-   * @remarks This method is oriented to AI Builders
-   * @remarks To call this method, the NVM API Key must have publication permissions
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/create-plan
-   *
-   * @param planMetadata - @see {@link PlanMetadata}
-   * @param priceConfig - @see {@link PlanPriceConfig}
-   * @param creditsConfig - @see {@link PlanCreditsConfig}
-   *
-   * @example
-   * ```
-   *  const freePriceConfig = getFreePriceConfig()
-   *  const 1dayDurationPlan = getExpirableDurationConfig(ONE_DAY_DURATION)
-   *  const { planId } = await payments.registerCreditsTrialPlan({name: 'Trial plan'}, freePriceConfig, 1dayDurationPlan)
-   * ```
-   *
-   * @returns The unique identifier of the plan (Plan ID) of the newly created plan.
-   */
-  public async registerCreditsTrialPlan(
-    planMetadata: PlanMetadata,
-    priceConfig: PlanPriceConfig,
-    creditsConfig: PlanCreditsConfig,
-  ): Promise<{ planId: string }> {
-    planMetadata.isTrialPlan = true
-    return this.registerCreditsPlan(planMetadata, priceConfig, creditsConfig)
-  }
-
-  /**
-   *
-   * It allows to an AI Builder to create a Trial Payment Plan on Nevermined limited by duration.
-   * A Nevermined Trial Plan allow subscribers of that plan to test the Agents associated to it.
-   * A Trial plan is a plan that only can be purchased once by a user.
-   * Trial plans, as regular plans, can be limited by duration (i.e 1 week of usage) or by credits (i.e 100 credits to use the agent).
-   * @remarks This method is oriented to AI Builders
-   * @remarks To call this method, the NVM API Key must have publication permissions
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/create-plan
-   *
-   * @param planMetadata - @see {@link PlanMetadata}
-   * @param priceConfig - @see {@link PlanPriceConfig}
-   * @param creditsConfig - @see {@link PlanCreditsConfig}
-   *
-   * @example
-   * ```
-   *  const freePriceConfig = getFreePriceConfig()
-   *  const 1dayDurationPlan = getExpirableDurationConfig(ONE_DAY_DURATION)
-   *  const { planId } = await payments.registerTimeTrialPlan({name: '1 day Trial plan'}, freePriceConfig, 1dayDurationPlan)
-   * ```
-   *
-   * @returns The unique identifier of the plan (Plan ID) of the newly created plan.
-   */
-  public async registerTimeTrialPlan(
-    planMetadata: PlanMetadata,
-    priceConfig: PlanPriceConfig,
-    creditsConfig: PlanCreditsConfig,
-  ): Promise<{ planId: string }> {
-    planMetadata.isTrialPlan = true
-    return this.registerTimePlan(planMetadata, priceConfig, creditsConfig)
-  }
-
-  /**
-   *
-   * It registers a new AI Agent on Nevermined.
-   * The agent must be associated to one or multiple Payment Plans. Users that are subscribers of a payment plan can query the agent.
-   * Depending on the Payment Plan and the configuration of the agent, the usage of the agent/service will consume credits.
-   * When the plan expires (because the time is over or the credits are consumed), the user needs to renew the plan to continue using the agent.
-   *
-   * @remarks This method is oriented to AI Builders
-   * @remarks To call this method, the NVM API Key must have publication permissions
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/register-agent
-   *
-   * @param agentMetadata - @see {@link AgentMetadata}
-   * @param agentApi - @see {@link AgentAPIAttributes}
-   * @param paymentPlans - the list of payment plans giving access to the agent.
-   *
-   * @example
-   * ```
-   *  const agentMetadata = { name: 'My AI Payments Agent', tags: ['test'] }
-   *  const agentApi = { endpoints: [{ 'POST': 'https://example.com/api/v1/agents/:agentId/tasks' }] }
-   *  const paymentPlans = [planId]
-   *
-   *  const { agentId } = await payments.registerAgent(agentMetadata, agentApi, paymentPlans)
-   * ```
-   *
-   * @returns The unique identifier of the newly created agent (Agent Id).
-   */
-  public async registerAgent(
-    agentMetadata: AgentMetadata,
-    agentApi: AgentAPIAttributes,
-    paymentPlans: string[],
-  ): Promise<{ agentId: string }> {
-    const body = {
-      metadataAttributes: agentMetadata,
-      agentApiAttributes: agentApi,
-      plans: paymentPlans,
-    }
-
-    const options = this.getBackendHTTPOptions('POST', body)
-    const url = new URL(API_URL_REGISTER_AGENT, this.environment.backend)
-
-    const response = await fetch(url, options)
-    if (!response.ok) {
-      throw new PaymentsError(
-        `Unable to register agent. ${response.statusText} - ${await response.text()}`,
-      )
-    }
-    const agentData = await response.json()
-    return { agentId: agentData.agentId }
-  }
-
-  /**
-   *
-   * It registers a new AI Agent and a Payment Plan associated to this new agent.
-   * Depending on the Payment Plan and the configuration of the agent, the usage of the agent/service will consume credits.
-   * When the plan expires (because the time is over or the credits are consumed), the user needs to renew the plan to continue using the agent.
-   *
-   * @remarks This method is oriented to AI Builders
-   * @remarks To call this method, the NVM API Key must have publication permissions
-   *
-   * @see https://docs.nevermined.app/docs/tutorials/builders/register-agent
-   *
-   * @param agentMetadata - @see {@link AgentMetadata}
-   * @param agentApi - @see {@link AgentAPIAttributes}
-   * @param planMetadata - @see {@link PlanMetadata}
-   * @param priceConfig - @see {@link PlanPriceConfig}
-   * @param creditsConfig - @see {@link PlanCreditsConfig}
-   *
-   * @example
-   * ```
-   *  const agentMetadata = { name: 'My AI Payments Agent', tags: ['test'] }
-   *  const agentApi { endpoints: [{ 'POST': 'https://example.com/api/v1/agents/:agentId/tasks' }] }
-   *  const cryptoPriceConfig = getNativeTokenPriceConfig(100n, builderAddress)
-   *  const 1dayDurationPlan = getExpirableDurationConfig(ONE_DAY_DURATION)
-   *  const { agentId, planId } = await payments.registerAgentAndPlan(
-   *    agentMetadata,
-   *    agentApi,
-   *    cryptoPriceConfig,
-   *    1dayDurationPlan
-   *  )
-   * ```
-   *
-   * @returns The unique identifier of the newly created agent (agentId).
-   * @returns The unique identifier of the newly created plan (planId).
-   */
-  public async registerAgentAndPlan(
-    agentMetadata: AgentMetadata,
-    agentApi: AgentAPIAttributes,
-    planMetadata: PlanMetadata,
-    priceConfig: PlanPriceConfig,
-    creditsConfig: PlanCreditsConfig,
-  ): Promise<{ agentId: string; planId: string }> {
-    const { planId } = await this.registerPlan(planMetadata, priceConfig, creditsConfig)
-    const { agentId } = await this.registerAgent(agentMetadata, agentApi, [planId])
-    return { agentId, planId }
-  }
-
-  /**
-   * Gets the metadata for a given Agent identifier.
-   *
-   * @param agentId - The unique identifier of the agent.
-   * @returns A promise that resolves to the agent's metadata.
-   * @throws PaymentsError if the agent is not found.
-   */
-  public async getAgent(agentId: string) {
-    const url = new URL(API_URL_GET_AGENT.replace(':agentId', agentId), this.environment.backend)
-    const response = await fetch(url)
-    if (!response.ok) {
-      throw new PaymentsError(`Agent not found. ${response.statusText} - ${await response.text()}`)
-    }
-    return response.json()
-  }
-
-  /**
-   * Gets the list of plans that can be ordered to get access to an agent.
-   *
-   * @param agentId - The unique identifier of the agent.
-   * @param pagination - Optional pagination options to control the number of results returned.p
-   * @returns A promise that resolves to the list of all different plans giving access to the agent.
-   * @throws PaymentsError if the agent is not found.
-   */
-  public async getAgentPlans(agentId: string, pagination = new PaginationOptions()) {
-    const query =
-      API_URL_GET_AGENT_PLANS.replace(':agentId', agentId) + '?' + pagination.asQueryParams()
-    const url = new URL(query, this.environment.backend)
-    console.log(`Fetching plans for agent ${agentId} from ${url.toString()}`)
-    const response = await fetch(url)
-    if (!response.ok) {
-      throw new PaymentsError(`Agent not found. ${response.statusText} - ${await response.text()}`)
-    }
-    return response.json()
-  }
-
-  /**
-   * Gets the information about a Payment Plan by its identifier.
-   *
-   * @param planId - The unique identifier of the plan.
-   * @returns A promise that resolves to the plan's description.
-   * @throws PaymentsError if the plan is not found.
-   */
-  public async getPlan(planId: string) {
-    const query = API_URL_GET_PLAN.replace(':planId', planId)
-    const url = new URL(query, this.environment.backend)
-    const response = await fetch(url)
-    if (!response.ok) {
-      throw new PaymentsError(`Plan not found. ${response.statusText} - ${await response.text()}`)
-    }
-    return response.json()
-  }
-
-  /**
-   * Gets the list of Agents that have associated a specific Payment Plan.
-   * All the agents returned can be accessed by the users that are subscribed to the Payment Plan.
-   *
-   * @param planId - The unique identifier of the plan.
-   * @param pagination - Optional pagination options to control the number of results returned.
-   * @returns A promise that resolves to the list of agents associated with the plan.
-   * @throws PaymentsError if the plan is not found.
-   */
-  public async getAgentsAssociatedToAPlan(planId: string, pagination = new PaginationOptions()) {
-    const query =
-      API_URL_GET_PLAN_AGENTS.replace(':planId', planId) + '?' + pagination.asQueryParams()
-    const url = new URL(query, this.environment.backend)
-    console.log(`Fetching agents for plan ${planId} from ${url.toString()}`)
-    const response = await fetch(url)
-    if (!response.ok) {
-      throw new PaymentsError(`Plan not found. ${response.statusText} - ${await response.text()}`)
-    }
-    return response.json()
-  }
-
-  /**
-   * Gets the balance of an account for a Payment Plan.
-   *
-   * @param planId - The identifier of the Payment Plan.
-   * @param accountAddress - The address of the account to get the balance for.
-   * @returns A promise that resolves to the balance result.
-   * @throws PaymentsError if unable to get the balance.
-   */
-  public async getPlanBalance(planId: string, accountAddress?: string): Promise<PlanBalance> {
-    const holderAddress = isEthereumAddress(accountAddress) ? accountAddress : this.accountAddress
-    const balanceUrl = API_URL_PLAN_BALANCE.replace(':planId', planId).replace(
-      ':holderAddress',
-      holderAddress!,
-    )
-
-    const options = {
-      method: 'GET',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-    }
-    const url = new URL(balanceUrl, this.environment.backend)
-    const response = await fetch(url, options)
-    if (!response.ok) {
-      throw new PaymentsError(
-        `Unable to get balance. ${response.statusText} - ${await response.text()}`,
-      )
-    }
-
-    return response.json()
+    return this.nvmApiKey.length > 0
   }
 
   /**
@@ -897,7 +428,7 @@ export class Payments {
    * @returns The information about the initialization of the request.
    * @throws PaymentsError if unable to initialize the agent request.
    */
-  public async initializeAgent(
+  public async startProcessingRequest(
     agentId: string,
     accessToken: string | undefined,
     urlRequested: string,
@@ -1049,24 +580,5 @@ export class Payments {
     }
 
     return response.json()
-  }
-
-  /**
-   * Returns the HTTP options required to query the backend.
-   * @param method - HTTP method.
-   * @param body - Optional request body.
-   * @returns HTTP options object.
-   * @internal
-   */
-  private getBackendHTTPOptions(method: string, body?: any) {
-    return {
-      method,
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${this.nvmApiKey}`,
-      },
-      ...(body && { body: JSON.stringify(body, jsonReplacer) }),
-    }
   }
 }
